Move some utility classes to HarpoS7.Utilities
