+++ HarpoS7.Tests/Auth/LegitimateSchemeTests.cs
@@ -1,5 +1,6 @@
 using System.Security.Cryptography;
 using HarpoS7.Auth;
+using HarpoS7.Utilities.Auth;
 using HarpoS7.Utilities.Extensions;
 
 namespace HarpoS7.Tests.Auth;
