+++ HarpoS7.Utilities/Auth/EPublicKeyFamily.cs
@@ -1,4 +1,4 @@
-namespace HarpoS7.Auth;
+namespace HarpoS7.Utilities.Auth;
 
 /// <summary>
 /// Public key family (read from the ObjectVariableTypeName item from CreateObject response)
