+++ HarpoS7/Auth/LegitimateScheme.cs
@@ -5,6 +5,7 @@ using HarpoS7.Aes;
 using HarpoS7.Utilities.Extensions;
 using HarpoS7.Keys;
 using HarpoS7.Seed;
+using HarpoS7.Utilities.Auth;
 
 namespace HarpoS7.Auth;
